Factor column_section_renderer smoketests into the new test_column_section_renderer.py
